+++ resources/views/partials/ui-config.blade.php
@@ -1,0 +1,5 @@
+@php
+    /** @var \LimenAi\Ui\ChatUiConfig $uiConfig */
+    $uiConfig = app(\LimenAi\Ui\ChatUiConfig::class);
+@endphp
+data-ui-config="{{ e($uiConfig->toJson()) }}"